+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/Screen.kt
@@ -2,8 +2,8 @@ package ai.mcpdirect.studio.app
 
 import ai.mcpdirect.studio.app.agent.MyStudioScreenDialog
 import ai.mcpdirect.studio.app.mcpkey.MCPKeyDialog
-import ai.mcpdirect.studio.app.model.account.AIPortAccessKey
 import ai.mcpdirect.studio.app.model.account.AIPortTeam
+import ai.mcpdirect.studio.app.model.aitool.AIPortToolAccessKey
 import ai.mcpdirect.studio.app.model.aitool.AIPortToolAgent
 import ai.mcpdirect.studio.app.model.aitool.AIPortToolMaker
 import ai.mcpdirect.studio.app.team.MCPTeamDialog
@@ -25,7 +25,7 @@ sealed class Screen(open val title: StringResource, open val icon: DrawableResou
 //    object ToolPermission : Screen(Res.string.tool_permission,
 //        Res.drawable.shield_toggle)
 
-    data class ToolPermission(val accessKey: AIPortAccessKey) : Screen(Res.string.tool_permission,
+    data class ToolPermission(val accessKey: AIPortToolAccessKey) : Screen(Res.string.tool_permission,
         Res.drawable.shield_toggle)
 
 //    object MyStudio : Screen(Res.string.connect_mcp,